Use void method rather than int
Originally, this method just determined the rank, and so having it return an `int`
made sense.

However, eventually, this method also came to set the instance field `tie`, which
it did not return, but merely set directly.

Accordingly, at that point the method became inconsistent in its type (that is, it was
both returning an `int` and setting an `int` behind the scenes).

As such, simply set both `int`s behind the scenes, and change the method to `void`.